Add restoreIndex field to store and init functions for wallet restore.
restoreIndex keeps track of how many seed words the user has entered so we know which ones are remaining for the user to enter.

We want to reset these values on the start of the restore process, as well as ensure they are up to date as the user goes from one page of restore entries to the next.